Fully Functional (With Bugs) + Added Pre-Heat
This version has all the final features built in and is working to
satisfactory standards. The only major bug I'm encountering is waking
from sleep. The unit will occasionally freeze when woken up. Other times
not.

Pre-Heat has been added. It is quite a robust feature and perhaps the
last feature that will fit without making the unit overly complicated.